Split core.cybercat into its own file
