+++ src/installers.ts
@@ -27,8 +27,6 @@ import {
   installCetCore,
   testForRedscriptCore,
   installRedscriptCore,
-  testForCyberCatCore,
-  installCoreCyberCat,
   testRed4ExtCore,
   installRed4ExtCore,
   testCoreCsvMerge,
@@ -55,6 +53,7 @@ import { makeSyntheticName } from "./installers.shared";
 import { extraFilesAllowedInOtherModTypesInstructions } from "./installer.special.extrafiles";
 import { InfoNotification, showInfoNotification } from "./ui.notifications";
 import { installCoreAmm, testForCoreAmm } from "./installer.core.amm";
+import { installCoreCyberCat, testForCyberCatCore } from "./installer.core.cybercat";
 import { installAmmMod, testForAmmMod } from "./installer.amm";
 
 // Ensure we're using win32 conventions
